backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pillow==10.1.0
numpy==1.24.3
opencv-python==4.9.0.80
torch==2.1.2
torchvision==0.16.2
diffusers==0.25.1
transformers==4.36.2
facenet-pytorch==2.5.0
insightface==0.7.3
onnxruntime==1.17.0
boto3==1.34.5
python-dotenv==1.0.0
aiofiles==23.2.1
requests==2.31.0
gunicorn==21.2.0
